Rename cct_parser -> file_parsers because it's used for multiple types now
