+++ gseim_grc/src/gseim/file_parsers/syntax_tree.py
